Update code to new memory mapping
I was having trouble keeping track of which registers and stack
variables were affected by which functions. To fix this, I decided to
rewrite the code to use statically-allocated variables for things like
the current instruction, the parameters extracted from the instruction,
as well as some scratch registers. This makes all the instruction
parsing and execution functions significantly more composable. It's now
very easy to ensure that only the work that needs to be done will be
done, and it's easy to re-use code between functions.

exec_op_imm is broken by this change, so it's disabled for now.