Fix OP-IMM decoding
Decoding SLLI and SRLI/SRAI clobbered A, so if they didn't match then no
other instructions that followed them in the decoding chain would be
properly decoded. To fix this, put funct3 in R3 instead and match on
that, then match on funct7 (in R7) only for the instructions that need
it.