Map SFRs into memory
This enables reading and writing arbitrary SFRs from RISC-V code.
Unfortunately, since SFRs need to be accessed directly, but we want to
be able to read an arbitrary one based on its address, we have to have a
big sequence of conditional jumps to pre-written SFR access
instructions, which nearly doubles the code size.

Maybe it would be possible to replace this with a jump table? And maybe
it's possible to do the same for the instruction decoders? At the very
least a jump table might be able to turn the current O(N) operation of
finding the matching value into an O(1) operation.

Also, a few short jumps had to be changed to long jumps since the
addition of the SFR access code increased the distance to the jump
targets outside the maximum range for short jump (and similar)
instructions.