Add test runner
This script will run the rv32ui-p tests from riscv-tests. The tests
themselves are not included yet because I have to make some
modifications to them to get them to work. The tests themselves are also
somewhat tricky to build, so I may just push the repo and then post the
compiled test binaries to that repo as part of a release.