Fix the LOAD/STORE tests
The .data section needs to be loaded into memory before the tests are
run. This can be done in the RISC-V test code, but it becomes
troublesome when you're trying to test the code you need to use to load
the data needed to test the code, so it's probably best to load the data
into memory using the simulator's memory write function instead.

With this, all tests are now passing.