Optimize arbitrary SFR byte reads
Switched from a chain of conditional jumps to a jump table. This means
there is now an 11-instruction delay before any move. It's a greater
delay than before for reading SFR addresses 0x80 to 0x89, but it's an
equal or shorter delay to access SFR addresses 0x8A-0xFF. It's also a
much shorter delay for accesses below 0x80, which always read as zero,
as it is now a constant 4-instruction delay instead of the
128-instruction delay it was before. In other words, not only has the
access time been reduced for most addresses, this means the time
complexity of an arbitrary SFR read operation has gone from O(n) to
approximately O(1). Finally, and probably most importantly, this change
reduces the size of the binary by 365 bytes, which is approximately a
10% savings.