Optimize arbitrary SFR byte writes
For the same benefits as in commit d12432f, switch to a jump table for
the SFR byte writes. This saves 324 bytes in the binary size--once again
a reduction of about 10%.

Also had to change the AJMP to exec_andi to an LJMP because the change
in binary size changed the page offset of the jump destination.
Fortunately, because exec_andi is the last entry in the jump table,
changing from a 2-byte AJMP to a 3-byte LJMP doesn't affect the offsets
of any of the other entries in the table.