Save the current instruction's DPTR before reading the instruction
This simplifies making calculations based on the address of the current
instruction, since it removes the need to subtract a hard-coded
instruction size from the next instruction's DPTR. This has the added
benefit of making it easier to support instructions that are not four
bytes in length (e.g., compressed instructions).

This change reduces the binary size by 4 bytes and speeds up AUIPC, JAL,
and conditional branch instructions by a net of 2 machine cycles. This
comes at the expense of a worst-case 4-cycle slowdown of all other
instructions.