Speed op decoding with binary search tree
This replaces serial tests for different op categories with a binary
tree based on bit tests. The particular bits were chosen by looking
for "groups" in the Karnaugh map. It produces about a 3% speedup in
the C "hello world" example, based on simulator ticks.

This adds three instructions and nine bytes.